Eraser and Paint Tweaks
Added erasing function with larger stroke width, because erasing with
the same width as painting is difficult af.
Fixed the painting outside of the canvas issue, Although the wider
eraser still covers the edge of the frame.
Fixed the progress bar drain to go down from full, rather than from full
to indeterminate.